Клиентская логика работы с групповыми чатами

+++ server/services/ChatService.cpp
@@ -3,6 +3,7 @@
 #include <drogon/HttpResponse.h>
 #include <json/forwards.h>
 #include <json/value.h>
+#include <algorithm>
 #include <cstddef>
 #include <cstdint>
 #include <optional>
@@ -16,6 +17,7 @@
 #include "jwt-cpp/jwt.h"
 #include "jwt-cpp/traits/kazuho-picojson/defaults.h"
 #include "models/Messages.h"
+#include "models/Users.h"
 #include "services/S3Service.hpp"
 #include "utils/Enum.hpp"
 #include "utils/server_exceptions.hpp"
@@ -30,6 +32,7 @@ using Message = drogon_model::messenger_db::Messages;
 using Chat = drogon_model::messenger_db::Chats;
 using ChatPreview = messenger::dto::ChatPreview;
 using Attachment = drogon_model::messenger_db::Attachments;
+using User = drogon_model::messenger_db::Users;
 
 Task<bool> ChatService::checkChatAccess(int64_t user_id, int64_t chat_id) {
     std::vector<messenger::repositories::ChatMember> chat_members =
@@ -63,6 +66,8 @@ Task<GetMessageByIdResponseDto> ChatService::getMessageById(
         throw messenger::exceptions::ForbiddenException("Access denied");
     }
 
+    std::optional<User> sender_info =
+        co_await user_repo->getById(message->getValueOfSenderId());
     std::vector<Attachment> attachments =
         co_await attachment_repo->getByMessage(message->getValueOfId());
     std::vector<std::optional<std::string>> attachments_download_urls(
@@ -79,7 +84,7 @@ Task<GetMessageByIdResponseDto> ChatService::getMessageById(
 
     co_return GetMessageByIdResponseDto(
         std::move(message.value()), std::move(attachments),
-        std::move(attachments_download_urls)
+        std::move(attachments_download_urls), std::move(sender_info)
     );
 }
 
@@ -93,34 +98,69 @@ Task<GetUserChatsResponseDto> ChatService::getUserChats(
         co_await chat_repo->getByUser(request_dto.from_request_user_id);
 
     std::vector<int64_t> message_ids;
+    std::unordered_set<int64_t> unique_senders_ids;
+
     for (const auto &chat_preview : chats_previews) {
         if (chat_preview.last_message.has_value()) {
             message_ids.push_back(chat_preview.last_message->getValueOfId());
+            unique_senders_ids.insert(
+                chat_preview.last_message->getValueOfSenderId()
+            );
         }
     }
 
+    std::vector<int64_t> senders_ids(
+        unique_senders_ids.begin(), unique_senders_ids.end()
+    );
+
     std::vector<std::vector<Attachment>> fetched_attachments =
         co_await attachment_repo->getByMessages(message_ids);
+
+    std::vector<User> fetched_senders_info;
+    if (!senders_ids.empty()) {
+        fetched_senders_info = co_await user_repo->getByIds(senders_ids);
+    }
 
     std::unordered_map<int64_t, std::vector<Attachment>> attachment_map;
     attachment_map.reserve(message_ids.size());
     for (std::size_t i = 0; i < message_ids.size(); i++) {
         attachment_map[message_ids[i]] = std::move(fetched_attachments[i]);
     }
+
+    std::unordered_map<int64_t, User> senders_map;
+    senders_map.reserve(fetched_senders_info.size());
+    for (auto &user : fetched_senders_info) {
+        senders_map[user.getValueOfId()] = std::move(user);
+    }
+
     std::vector<std::vector<Attachment>> last_messages_attachments;
+    std::vector<std::optional<User>> last_messages_senders;
     last_messages_attachments.reserve(chats_previews.size());
+    last_messages_senders.reserve(chats_previews.size());
 
     for (const auto &chat_preview : chats_previews) {
         if (chat_preview.last_message.has_value()) {
             int64_t msg_id = chat_preview.last_message->getValueOfId();
+            int64_t sender_id = chat_preview.last_message->getValueOfSenderId();
+
             last_messages_attachments.push_back(std::move(attachment_map[msg_id]
             ));
+
+            auto it = senders_map.find(sender_id);
+            if (it != senders_map.end()) {
+                last_messages_senders.push_back(it->second);
+            } else {
+                last_messages_senders.push_back(std::nullopt);
+            }
         } else {
             last_messages_attachments.push_back({});
+            last_messages_senders.push_back(std::nullopt);
         }
     }
+
     GetUserChatsResponseDto response_dto(
-        std::move(chats_previews), std::move(last_messages_attachments)
+        std::move(chats_previews), std::move(last_messages_attachments),
+        std::move(last_messages_senders)
     );
     co_return response_dto;
 }
@@ -182,9 +222,32 @@ Task<GetChatMessagesResponseDto> ChatService::getChatMessages(
             );
         }
     }
+
+    std::unordered_set<int64_t> unique_sender_ids;
+    for (const auto &message : chat_messages) {
+        unique_sender_ids.insert(message.getValueOfSenderId());
+    }
+    std::vector<int64_t> senders_ids(
+        unique_sender_ids.begin(), unique_sender_ids.end()
+    );
+
+    std::vector<User> fetched_senders =
+        co_await user_repo->getByIds(senders_ids);
+
+    std::unordered_map<int64_t, User> senders_map;
+    for (auto &user : fetched_senders) {
+        senders_map[user.getValueOfId()] = std::move(user);
+    }
+
+    std::vector<User> senders_info;
+    senders_info.reserve(chat_messages.size());
+    for (const auto &message : chat_messages) {
+        senders_info.push_back(senders_map[message.getValueOfSenderId()]);
+    }
+
     GetChatMessagesResponseDto response_dto(
         std::move(chat_messages), std::move(attachments),
-        std::move(attachments_download_urls)
+        std::move(attachments_download_urls), std::move(senders_info)
     );
     co_return response_dto;
 }
@@ -396,12 +459,15 @@ Task<GetAttachmentLinksResponseDto> ChatService::getAttachmentLinks(
 Task<CreateGroupResponseDto> ChatService::createGroup(
     CreateGroupRequestDto request_dto
 ) {
-    for (auto member_id : request_dto.members_ids) {
-        if (member_id == request_dto.creator_id) {
-            throw messenger::exceptions::BadRequestException(
-                "There is creator id in members_ids list"
-            );
-        }
+    auto it = std::find(
+        request_dto.members_ids.begin(), request_dto.members_ids.end(),
+        request_dto.creator_id
+    );
+
+    if (it == request_dto.members_ids.end()) {
+        throw messenger::exceptions::BadRequestException(
+            "Creator ID must be included in the members list"
+        );
     }
     Chat chat = co_await chat_repo->createGroup(
         request_dto.name, request_dto.creator_id,
@@ -422,7 +488,10 @@ Task<AddGroupChatMemberResponseDto> ChatService::addGroupChatMember(
             "Not enough permissions to add members to this chat"
         );
     }
-    auto chat_members = co_await chat_repo->getMembers(request_dto.chat_id);
+    auto trans_ptr = co_await drogon::app().getDbClient()->newTransactionCoro();
+    co_await chat_repo->lockChat(request_dto.chat_id, trans_ptr);
+    auto chat_members =
+        co_await chat_repo->getMembers(request_dto.chat_id, trans_ptr);
     if (chat_members.size() >= 50) {
         throw messenger::exceptions::ConflictException(
             "There are already 50 members in this chat"
@@ -433,8 +502,10 @@ Task<AddGroupChatMemberResponseDto> ChatService::addGroupChatMember(
         );
     }
     auto new_member = co_await chat_repo->addMember(
-        request_dto.chat_id, request_dto.new_member_id, request_dto.role
+        request_dto.chat_id, request_dto.new_member_id, request_dto.role,
+        trans_ptr
     );
+    co_await trans_ptr->execSqlCoro("COMMIT;");
     AddGroupChatMemberResponseDto response_dto(std::move(new_member));
     co_return response_dto;
 }
@@ -452,7 +523,15 @@ Task<GetChatMemberResponseDto> ChatService::getChatMember(
     auto member = co_await chat_repo->getMember(
         request_dto.chat_id, request_dto.member_id
     );
-    GetChatMemberResponseDto response_dto(std::move(member));
+    auto member_info = co_await user_repo->getById(member.getValueOfUserId());
+    if (!member_info.has_value()) {
+        throw messenger::exceptions::NotFoundException(
+            "Couldn't find user with member's user_id"
+        );
+    }
+    GetChatMemberResponseDto response_dto(
+        std::move(member), std::move(member_info.value())
+    );
     co_return response_dto;
 }
 
@@ -467,7 +546,32 @@ Task<GetChatMembersResponseDto> ChatService::getChatMembers(
         );
     }
     auto members = co_await chat_repo->getMembers(request_dto.chat_id);
-    GetChatMembersResponseDto response_dto(std::move(members));
+    std::vector<int64_t> members_user_ids;
+    members_user_ids.reserve(members.size());
+    for (auto &el : members) {
+        members_user_ids.push_back(el.getValueOfUserId());
+    }
+    std::vector<User> fetched_members_info =
+        co_await user_repo->getByIds(members_user_ids);
+
+    std::unordered_map<int64_t, User> users_map;
+    for (auto &user : fetched_members_info) {
+        users_map[user.getValueOfId()] = std::move(user);
+    }
+    std::vector<User> aligned_members_info;
+    aligned_members_info.reserve(members.size());
+
+    for (const auto &member : members) {
+        auto it = users_map.find(member.getValueOfUserId());
+        if (it != users_map.end()) {
+            aligned_members_info.push_back(it->second);
+        } else {
+            aligned_members_info.push_back(User{});
+        }
+    }
+    GetChatMembersResponseDto response_dto(
+        std::move(members), std::move(aligned_members_info)
+    );
     co_return response_dto;
 }
 
@@ -490,6 +594,24 @@ Task<RemoveMemberResponseDto> ChatService::removeMember(
         members.size() > 1) {
         throw messenger::exceptions::ConflictException(
             "Owner cannot leave chat without transfer of rights"
+        );
+    }
+    auto it = std::find_if(members.begin(), members.end(), [&](const auto &m) {
+        return m.getValueOfUserId() == request_dto.member_id;
+    });
+    if (it == members.end()) {
+        throw messenger::exceptions::NotFoundException(
+            "Member not found in chat"
+        );
+    }
+    auto to_delete_member = *it;
+    if (request_source_member.getValueOfRole() ==
+            messenger::models::ChatRole::Admin &&
+        request_dto.user_id != request_dto.member_id &&
+        to_delete_member.getValueOfRole() !=
+            messenger::models::ChatRole::Member) {
+        throw messenger::exceptions::ForbiddenException(
+            "Admins can only remove regular members"
         );
     }
     co_await chat_repo->removeMember(
@@ -547,4 +669,19 @@ Task<UpdateChatInfoResponseDto> ChatService::updateChatInfo(
         request_dto.description
     );
     co_return UpdateChatInfoResponseDto();
+}
+
+Task<GetChatByIdResponseDto> ChatService::getChatById(
+    GetChatByIdRequestDto request_dto
+) {
+    // TODO: access checking if chat is private
+    auto chat = co_await chat_repo->getById(request_dto.chat_id);
+    if (!chat.has_value()) {
+        throw messenger::exceptions::NotFoundException(
+            "Chat with id " + std::to_string(request_dto.chat_id) +
+            " doesn't exist"
+        );
+    }
+    GetChatByIdResponseDto response_dto(std::move(chat.value()));
+    co_return response_dto;
 }
